[refs #91506] WIP plone utilities for RabbitMQ client as extra feature.
